move ncdf4 from Imports to Suggests
The ncdf4 package requires installing the NetCDF library on Linux
systems before package install. This is unnecessary, tedious, and
confusing for the majority of users who will use GDS files. Moving
the ncdf4 requirement to Suggests streamlines the install process
for the majority of users of this package and other packages that
depend on it.